test: fix test target compile error and stale MLX dimension expectation
- SearchEngineTests: `hash &+ i * 31` mixed UInt64 and Int and did not
  compile with Swift 6.3.3; promote the index to UInt64.
- EmbeddingProviderTests: MLXEmbeddingProvider() has defaulted to
  Qwen3-Embedding-0.6B (1024 dims) since January; the registry test
  still expected 768.

Both were hidden by the CI pipeline swallowing swift's exit status.

+++ Tests/SwiftIndexCoreTests/SearchEngineTests.swift
@@ -244,7 +244,7 @@ actor SearchTestMockEmbeddingProvider: EmbeddingProvider {
         var vector = [Float](repeating: 0, count: dimension)
         let hash = stableTestHash(text)
         for i in 0 ..< dimension {
-            vector[i] = Float((hash &+ i * 31) % 1000) / 1000.0
+            vector[i] = Float((hash &+ UInt64(i) * 31) % 1000) / 1000.0
         }
         return vector
     }
